requirements.txt
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.4
openai==1.68.2
pandas==2.2.3
pt_core_news_md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.8.0/pt_core_news_md-3.8.0-py3-none-any.whl#sha256=54382cda034e41f3ec605ceeb924cd6cfdbced00a65f7afa12218520ba7008c5
scikit-learn==1.6.1
scipy==1.15.2
spacy==3.8.4
wordcloud==1.9.4
pyyaml~=6.0.2